requirements.txt
apache-airflow
tweepy
textblob
nltk
pandas
